requirements.txt
pika==1.0.0
protobuf==3.7.1
mysql-connector==2.2.9
